Move profile buttons to top of page, add skills description
