+++ components/MySkills/MySkills.css
@@ -18,3 +18,25 @@
 .tech-section-fade {
     background-image: linear-gradient(to bottom, #000000, var(--tech-background-color));
 }
+.skills-description {
+    height: 1.5em;
+    padding: 0 10px;
+    font-style: italic;
+    opacity: 0.7;
+    font-size: 0.9em;
+}
+@media(min-width: 653px) and (max-width: 1169px) {
+    .skills-description {
+        height: 2.5em;
+    }
+}
+@media (min-width: 501px) and (max-width: 652px) {
+    .skills-description {
+        height: 3.5em;
+    }
+}
+@media (max-width: 500px) {
+    .skills-description {
+        height: inherit;
+    }
+}
