+++ components/SelfIntro/SelfIntro.js
@@ -24,6 +24,11 @@ class SelfIntro extends HTMLElement {
 	            </div>
 	            <div class="intro-text-spacer"></div>
 	        </div>
+	        <div class="profiles">
+                <a class="profile-btn" href="https://github.com/robere2" aria-label="GitHub"><i class="fab fa-github"></i></a>
+                <a class="profile-btn" href="https://github.com/robere2" aria-label="LinkedIn"><i class="fab fa-linkedin"></i></a>
+                <a class="profile-btn" href="https://www.npmjs.com/~bugfroggy" aria-label="NPM"><i class="fab fa-npm"></i></a>
+			</div>
 		`;
 	}
 }
